Added Logic for Speed Warning Retract Flaps
Ideally retraction should depend on Indicated Air Speed, but I could not find it. Used trueAirSpeed instead.
Also the ranges for flap extraction and max speed allowed for Airbus A350 and B737 vary slightly, therefore used FlapRatio instead.
Note that in the B737, the 8 increments of flap settings DO Not have the same step, unlike the 4 increments in Airbus.
Audio is generated by AI, as I could not find a suitable audio clip.